requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
flake8==5.0.4
idna==3.4
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
types-requests==2.28.11.2
types-urllib3==1.26.25.1
typing_extensions==4.4.0
urllib3==1.26.12
